refactor: useProducts 훅 분리
useProducts 내 기능(query얻기, fetch)를 훅으로 분리
